# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Syra API",
    "version": "1.0.0",
    "contact": {
      "email": "support@syraa.fun"
    },
    "x-guidance": "Free routes (security: []) include /info and /prediction-game/health. Paid routes return HTTP 402 until settled via x402 (Solana/Base USDC) — pay an offer from the 402 body accepts array and retry with PAYMENT-SIGNATURE or X-PAYMENT. Docs: https://docs.syraa.fun",
...
  },
  "servers": [
    {
      "url": "https://api.syraa.fun"
    }
  ],
...
  "paths": {
    "/signal": {
      "get": {
        "tags": [
          "Market data (x402)"
        ],
        "summary": "AI trading signal (x402)",
        "description": "Spot OHLC + technical signal engine. Default source: binance with CEX fallbacks (coinbase, okx, kraken). Use source=n8n|webhook for legacy webhook.",
        "operationId": "getSignal",
        "security": [
          {
            "x402": []
          }
        ],
        "x-payment-info": {
          "protocols": [
            {
              "x402": {}
            },
            {
              "mpp": {
                "method": "",
                "intent": "",
                "currency": ""
              }
            }
          ],
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.1"
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "bitcoin"
            },
            "description": "Asset name for signal routing (e.g. bitcoin, solana)."
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Data source. Default: binance. CEX: binance, coinbase, coingecko, okx, bybit, kraken, bitget, kucoin, upbit, cryptocom (alias: crypto.com → cryptocom). n8n | webhook for legacy webhook."
          },
          {
            "name": "instId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Venue symbol override (e.g. BTCUSDT, BTC-USDT)."
          },
          {
            "name": "bar",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Candle interval (e.g. 1m, 1h, 4h, 1d)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Candle count (default ~200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Signal result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "signal": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "402": {
            "description": "Payment required — x402 v2 (Payment-Required header + JSON body with accepts). Retry with PAYMENT-SIGNATURE or X-PAYMENT.",
            "headers": {
              "Payment-Required": {
                "description": "x402 v2 encoded payment requirements",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests — burst (25 req / 10s per IP) or sustained (100 req / 60s per IP) limit exceeded. Wait for the duration in `Retry-After` (seconds) before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Seconds the client must wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Too many requests. Please slow down."
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Upstream or configuration error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Market data (x402)"
        ],
        "summary": "AI trading signal via JSON body (x402)",
        "operationId": "postSignal",
        "security": [
          {
            "x402": []
          }
        ],
        "x-payment-info": {
          "protocols": [
            {
              "x402": {}
            },
            {
              "mpp": {
                "method": "",
                "intent": "",
                "currency": ""
              }
            }
          ],
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.1"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "source": {
                    "type": "string"
                  },
                  "instId": {
                    "type": "string"
                  },
                  "bar": {
                    "type": "string"
                  },
                  "limit": {
                    "type": "number"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signal result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "signal": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "402": {
            "description": "Payment required — x402 v2 (Payment-Required header + JSON body with accepts). Retry with PAYMENT-SIGNATURE or X-PAYMENT.",
            "headers": {
              "Payment-Required": {
                "description": "x402 v2 encoded payment requirements",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests — burst (25 req / 10s per IP) or sustained (100 req / 60s per IP) limit exceeded. Wait for the duration in `Retry-After` (seconds) before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Seconds the client must wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Too many requests. Please slow down."
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "PAYMENT-SIGNATURE",
        "description": "x402 v2 payment proof (also accepts X-PAYMENT header). Obtain by paying an offer from the HTTP 402 response."
      },
      "SyraApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Legacy internal routes only — openapi-listed free and x402 routes do not use this."
      }
    }
  }
}